notificate-documents-deleted
