+++ .gitignore
@@ -10,6 +10,7 @@ capybara-*.html
 /public/system
 /coverage/
 /spec/tmp
+/public/uploads
 *.orig
 rerun.txt
 pickle-email-*.html
