+++ app/controllers/documents_controller.rb
@@ -1,7 +1,6 @@
 class DocumentsController < ApplicationController
   before_action :logged_in_user, only: %i(new create destroy)
   before_action :find_document, only: %i(destroy show)
-  before_action :correct_document, only: :destroy
   before_action :load_data_comment, only: :show
   before_action :add_to_history_view, only: :show
