+++ app/controllers/search_documents_controller.rb
@@ -1,5 +1,6 @@
 class SearchDocumentsController < ApplicationController
   before_action :checked_admin, only: :search_reported
+  before_action :correct_user, only: :search_deleted
 
   def search_name
     @documents = Document.search_name(params[:document][:name_document])
@@ -15,5 +16,10 @@ class SearchDocumentsController < ApplicationController
   def search_reported
     @reports = Comment.status_report(true).group_by(&:document_id)
     @documents = Document.search_id(@reports.keys).paginate page: params[:page]
+  end
+
+  def search_deleted
+    user = User.find_by id: params[:id]
+    @documents = user.documents.only_deleted.paginate page: params[:page]
   end
 end
