+++ app/views/documents/_document.html.erb
@@ -1,15 +1,21 @@
 <li id="document-<%= document.id %>">
   <span class="content">
-    <h3 class="text-info"> <%= document.name_document %></h3>
+    <h4 class="text-info"> <%= document.name_document %></h4>
     <%= t ".upload_by", name: document.user.name %>
   </span>
-  <p class="content"> <%= t ".download_times", times: document.history_downloads.size %></p>
   <span class="timestamp">
+    <%= t ".download_times", times: document.history_downloads.size %></br>
     <%= t ".posted", time: time_ago_in_words(document.created_at) %>
-    <% if current_user? document.user %>
-      <%= link_to t(".del"), document, method: :delete, data: {confirm: t(".confirm")} %>
+    <% unless document.deleted_at? %>
+      <% if current_user? document.user %>
+        | <%= link_to document, method: :delete, data: {confirm: t(".confirm")} do %>
+          <span class="glyphicon glyphicon-trash"></span>
+        <% end %>
+      <% end %> </br>
+      <%= link_to t(".view_detail") , document_path(document.id) %>
+      | <%= link_to document.content.url do %>
+        <span class="glyphicon glyphicon-eye-open"></span>
+      <% end %>
     <% end %>
   </span>
-  <%= link_to t(".view_detail") , document_path(document.id) %>
-  <%= link_to t(".open_document") , document.content.url %>
 </li>
