+++ app/views/shared/_user_info.html.erb
@@ -47,5 +47,12 @@
         <%= @user.coin %>
       </strong>
     <% end %>
+
+    <%= link_to search_deleted_url(id: @user.id) do %>
+      <%= t ".document_delete" %>
+      <strong id="document_delete" class="stat text-center">
+        <%= @user.documents.only_deleted.size %>
+      </strong>
+    <% end %>
   </div>
 </section>
